backend/requirements.txt
APScheduler==3.7.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
future==0.18.2
greenlet==1.0.0
h11==0.12.0
idna==2.10
line-bot-sdk==1.19.0
paho-mqtt==1.5.1
psycopg2-binary==2.8.6
pydantic==1.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.0
starlette==0.13.6
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.13.4
